+++ pkg/kt_observability_monitoring/http_client_metrics.go
@@ -6,8 +6,9 @@ import (
 
 // If you develop a client for a HTTP API you can use this class to quickly and efficiently attach Metrics to your client.
 //
-// This object is designed the way that it starts empty when created (has 0 Metric) and Metrics are getting created and exposed as you invoke it's methods. This is why it is "lazy".
-// You can track how many times requests were sent, and how many times they succeeded / failed. You also have the possibility to track Request-Response loop times AND you can do it
+// This object is designed the way that it starts empty when created (has 0 Metric) and Metrics are getting created and exposed as you invoke it's methods. This
+// is why it is "lazy". You can track how many times requests were sent, and how many times they succeeded / failed. You also have the possibility to track
+// Request-Response loop times AND you can do it
 // per each HttpSatatus codes if you want which brings pretty good observability just out of the box.
 type HttpClientLazyMetricsSet struct {
 	of        string
@@ -47,7 +48,8 @@ func NewHttpClientLazyMetricsSet(of string, opts ...HttpClientLazyMetricsSetOpt)
 	return &metrics
 }
 
-// Assigns a "qualifier" to all Metric instances in your set of your choice. One example of good qualifiers could be the httpMethod like GET, POST, PUT etc to reflect request type - but can be anything else too of course.
+// Assigns a "qualifier" to all Metric instances in your set of your choice. One example of good qualifiers could be the httpMethod like GET, POST, PUT etc to
+// reflect request type - but can be anything else too of course.
 func WithHttpClientQualifier(qualifier any) HttpClientLazyMetricsSetOpt {
 	return func(m *HttpClientLazyMetricsSet) {
 		if qualifier != nil {
@@ -78,40 +80,55 @@ func WithClientId(id string) HttpClientLazyMetricsSetOpt {
 // Invoke when client sent the request - will create+increase counter
 func (m *HttpClientLazyMetricsSet) RequestSent() {
 	if m.reqSentCounter == nil {
-		c := GetCounterMetricInstance(GetClientRequestSentCountTemplate(), map[string]interface{}{"of": m.of, "protocol": "http", "statusCode": "-", "qualifier": m.qualifier, "clientId": m.clientId})
+		c := GetCounterMetricInstance(
+			GetClientRequestSentCountTemplate(),
+			map[string]any{"of": m.of, "protocol": "http", "statusCode": "-", "qualifier": m.qualifier, "clientId": m.clientId},
+		)
 		m.reqSentCounter = &c
 	}
 	(*m.reqSentCounter).Inc()
 }
 
 // Invoke when client received a success - pass in the httpStatusCode what was returned. This will create+increase the appropriate success counter.
-// The statusCode is taken as a string although normally it is int. Reason: this way if you do not want to distinguish fully just by ranges let's say you can send "2xx" to represent anything in 2xx range.
+// The statusCode is taken as a string although normally it is int. Reason: this way if you do not want to distinguish fully just by ranges let's say you can
+// send "2xx" to represent anything in 2xx range.
 func (m *HttpClientLazyMetricsSet) RequestSucceeded(withHttpStatusCode string) {
 	c, found := m.reqSuccessCounterByStatusCode[withHttpStatusCode]
 	if !found {
-		c = GetCounterMetricInstance(GetClientRequestSucceededCountTemplate(), map[string]interface{}{"of": m.of, "protocol": "http", "statusCode": withHttpStatusCode, "qualifier": m.qualifier, "clientId": m.clientId})
+		c = GetCounterMetricInstance(
+			GetClientRequestSucceededCountTemplate(),
+			map[string]any{"of": m.of, "protocol": "http", "statusCode": withHttpStatusCode, "qualifier": m.qualifier, "clientId": m.clientId},
+		)
 		m.reqSuccessCounterByStatusCode[withHttpStatusCode] = c
 	}
 	c.Inc()
 }
 
 // Invoke when client received a failure - pass in the httpStatusCode of the failure. This will create+increase the appropriate failure counter
-// The statusCode is taken as a string although normally it is int. Reason: this way if you do not want to distinguish fully just by ranges let's say you can send "5xx" to represent anything in 5xx range.
+// The statusCode is taken as a string although normally it is int. Reason: this way if you do not want to distinguish fully just by ranges let's say you can
+// send "5xx" to represent anything in 5xx range.
 func (m *HttpClientLazyMetricsSet) RequestFailed(withHttpStatusCode string) {
 	c, found := m.reqFailedCounterByStatusCode[withHttpStatusCode]
 	if !found {
-		c = GetCounterMetricInstance(GetClientRequestFailedCountTemplate(), map[string]interface{}{"of": m.of, "protocol": "http", "statusCode": withHttpStatusCode, "qualifier": m.qualifier, "clientId": m.clientId})
+		c = GetCounterMetricInstance(
+			GetClientRequestFailedCountTemplate(),
+			map[string]any{"of": m.of, "protocol": "http", "statusCode": withHttpStatusCode, "qualifier": m.qualifier, "clientId": m.clientId},
+		)
 		m.reqFailedCounterByStatusCode[withHttpStatusCode] = c
 	}
 	c.Inc()
 }
 
 // Track processing times - pass in the httpStatusCode so we can collect segregated. This will maintain a Summary
-// The statusCode is taken as a string although normally it is int. Reason: this way if you do not want to distinguish fully just by ranges let's say you can send "2xx" to represent anything in 2xx range.
+// The statusCode is taken as a string although normally it is int. Reason: this way if you do not want to distinguish fully just by ranges let's say you can
+// send "2xx" to represent anything in 2xx range.
 func (m *HttpClientLazyMetricsSet) RequestTookMillis(httpStatusCode string, millis float64) {
 	c, found := m.reqProcessingTimeByStatusCode[httpStatusCode]
 	if !found {
-		c = GetSummaryMetricInstance(GetClientRequestProcessingTimeTemplate(), map[string]interface{}{"of": m.of, "protocol": "http", "statusCode": httpStatusCode, "qualifier": m.qualifier, "clientId": m.clientId})
+		c = GetSummaryMetricInstance(
+			GetClientRequestProcessingTimeTemplate(),
+			map[string]any{"of": m.of, "protocol": "http", "statusCode": httpStatusCode, "qualifier": m.qualifier, "clientId": m.clientId},
+		)
 		m.reqProcessingTimeByStatusCode[httpStatusCode] = c
 	}
 	c.Observe(millis)